+++ salem/sio.py
@@ -395,7 +395,7 @@ class _XarrayAccessorBase(object):
         dim = utils.str_in_list(dn, utils.valid_names['z_dim'])
         self.z_dim = dim[0] if dim else None
 
-    def subset(self, margin=0, **kwargs):
+    def subset(self, margin=0, ds=None, **kwargs):
         """subset(self, margin=0, shape=None, geometry=None, grid=None,
                   corners=None, crs=wgs84, roi=None)
 
@@ -405,6 +405,8 @@ class _XarrayAccessorBase(object):
 
         Parameters
         ----------
+        ds : Dataset or DataArray
+            form the ROI from the extent of the Dataset or DataArray
         shape : str
             path to a shapefile
         geometry : geometry
@@ -424,6 +426,10 @@ class _XarrayAccessorBase(object):
             five pixels from each boundary of the dataset.
         """
 
+        if ds is not None:
+            grid = ds.salem.grid
+            kwargs.setdefault('grid', grid)
+
         mask = self.grid.region_of_interest(**kwargs)
         if not kwargs:
             # user just wants a margin
@@ -438,7 +444,7 @@ class _XarrayAccessorBase(object):
                           ]
         return out_ds
 
-    def roi(self, **kwargs):
+    def roi(self, ds=None, **kwargs):
         """roi(self, shape=None, geometry=None, grid=None, corners=None,
                crs=wgs84, roi=None)
 
@@ -448,6 +454,8 @@ class _XarrayAccessorBase(object):
 
         Parameters
         ----------
+        ds : Dataset or DataArray
+            form the ROI from the extent of the Dataset or DataArray
         shape : str
             path to a shapefile
         geometry : geometry
@@ -462,6 +470,10 @@ class _XarrayAccessorBase(object):
         roi : ndarray
             if you have a mask ready, you can give it here
         """
+
+        if ds is not None:
+            grid = ds.salem.grid
+            kwargs.setdefault('grid', grid)
 
         mask = self.grid.region_of_interest(**kwargs)
         coords = {self.y_dim: self._obj[self.y_dim].values,